+++ packages/components/src/search-control/user-search-control.js
@@ -31,16 +31,16 @@ export default function UserSearchControl( {
 	formatSearchResultLabel = formatSearchResultLabel
 		? formatSearchResultLabel
 		: ( { name, id } ) =>
-				sprintf(
-					// Translators: %1$s = User's name; %2$s = User's id.
-					_x(
-						'%1$s (ID# %2$d)',
-						'User search result label',
-						'lifterlms'
-					),
-					name,
-					id
-				);
+			sprintf(
+				// Translators: %1$s = User's name; %2$s = User's id.
+				_x(
+					'%1$s (ID# %2$d)',
+					'User search result label',
+					'lifterlms'
+				),
+				name,
+				id
+			);
 
 	return (
 		<BaseSearchControl