+++ packages/components/src/spinner/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+	env: {
+		browser: true,
+	},
+	globals: {
+		jQuery: true,
+	},
+};