+++ packages/icons/docs/app.js
@@ -1,16 +1,19 @@
+import { render } from 'react-dom';
+import * as Icons from '../src';
+
+const { Icon, ...icons } = Icons;
+
 /**
  * React App that automatically renders all icons in the library into a static file.
  *
  * This app is rendered as a static HTML file and generate.js pulls the HTML the icons
  * and stores them as raw SVG files for display use in the README.md file between
  * the appropriate comment tokens.
+ *
+ * @since 1.0.0
+ *
+ * @return {Object} React component fragment.
  */
-
-import { render } from 'react-dom';
-import * as Icons from '../src';
-
-const { Icon, ...icons } = Icons;
-
 function App() {
 	return (
 		<>